+++ src/main/scala/whu/edu/cn/geocube/core/raster/ingest/OlapIngestor.scala
@@ -1,10 +1,5 @@
 package whu.edu.cn.geocube.core.raster.ingest
 
-import java.io.File
-import java.sql.{ResultSet, SQLException}
-import whu.edu.cn.geocube.util.PostgresqlService
-import whu.edu.cn.util.PostgresqlUtil
-
 /**
  * Save the results of on-the-fly computation to database for precomputation.
  */
